Updated liveview to send state to the input field for disabling
